feat: reorganize screen packages and update imports for drawing functionality
